run processFile() in parallel, using a workerpool
