+++ lib/worker.js
@@ -1,0 +1,15 @@
+'use strict';
+
+var workerpool = require('workerpool');
+
+var processFile = require('./process-file');
+
+// TODO - with an option to disable this parallelism
+function processFileParallel(inFile, outFile, relativePath, outDir, silent, _options) {
+  return processFile(inFile, outFile, relativePath, outDir, silent, _options);
+}
+
+// create worker and register public functions
+workerpool.worker({
+  processFileParallel
+});
