+++ package.json
@@ -9,7 +9,8 @@
   "license": "MIT",
   "author": "Edward Faulkner <ef@alum.mit.edu>",
   "files": [
-    "index.js"
+    "index.js",
+    "lib"
   ],
   "main": "index.js",
   "repository": {
@@ -22,6 +23,7 @@
     "test:watch": "jest --watchAll"
   },
   "dependencies": {
+    "async-promise-queue": "^1.0.4",
     "broccoli-plugin": "^1.2.1",
     "debug": "^3.1.0",
     "lodash.defaultsdeep": "^4.6.0",
@@ -30,7 +32,8 @@
     "source-map-url": "^0.4.0",
     "symlink-or-copy": "^1.0.1",
     "uglify-es": "^3.1.3",
-    "walk-sync": "^0.3.2"
+    "walk-sync": "^0.3.2",
+    "workerpool": "^2.3.0"
   },
   "devDependencies": {
     "babel-jest": "^21.2.0",
@@ -53,6 +56,7 @@
     "modulePathIgnorePatterns": [
       "<rootDir>/tmp"
     ],
+    "testEnvironment": "node",
     "testMatch": [
       "<rootDir>/test/test.js"
     ],
